update gitignore, body parser, create config, refactoring
