+++ .gitignore
@@ -1,1 +1,4 @@
 node_modules
+config
+package-lock.json
+yarn.lock
