+++ routes/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { ensureAuthenticated } = require('../config/auth');
+const { ensureAuthenticated } = require('../middlewares/auth');
 
 // Welcome Page
 router.get('/', (req, res) => res.render('welcome'));
